pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "window-translation"
version = "0.1.0"
description = "Capture a region of the screen, OCR it, translate with an AI API, and show the result as an overlay."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "window_translation contributors" }]
keywords = ["ocr", "translation", "overlay", "screen-capture", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: X11 Applications :: Qt",
    "Environment :: Win32 (MS Windows)",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Multimedia :: Graphics :: Capture :: Screen Capture",
]
dependencies = [
    "PySide6>=6.6",
    "mss>=9.0",
    "Pillow>=11.0",
    "pytesseract>=0.3.10",
    "pynput>=1.7.6",
    "requests>=2.32.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-qt>=4.2",
]

[project.scripts]
window-translation = "window_translation.app:main"

[project.gui-scripts]
window-translation-gui = "window_translation.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
